Use ff-oracle-jre-installer to install JRE
This new utility was written to install JRE and provide
a feedback to user when installation is complete.

That's the continuation of the approach of installing JRE
on flatpak run phase, due to the custom cookie we have
to provide in order to download JRE from Oracle. Unfortunately,
Flatpak's extra_data functionality isn't able to use custom
cookies.