Use Firefox's binaries from extra-data instead of compiling them
This is needed in order to comply with the Mozilla trademark's terms
of use. This also means that most of the compiled sources were removed,
making the build much faster.

Because now Firefox's executable will be placed in a different path
(/app/extra/...), this patch also changes the command from
'firefox-wrapper' to simply 'firefox'.

https://phabricator.endlessm.com/T14332