Fix installing the plugins
We install the Flash and JAVA plugins together with Firefox, however
the way this was done had a few issues:
- For Flash, it was being done through the extra-data, which means that
  the same version of Flash would always be downloaded on every update.
  The plugin should be changed only if there are new versions, so
  downloading and installing it from a script shipped in the Flatpak app
  is a better approach.
- For JAVA, the Python script that was installing it did not work.
  Besides, a few Python dependencies were being installed (some of them
  not really needed, like a CLI progress bar).

This patch replaces the logic for both plugins by a single bash script
that is run together with Firefox and installs the plugins when needed.
It also installs them in the XDG_DATA_HOME which is better than in the
user's home (if the user denied access to their home to the app, the
plugins would not be found, so a new update would be performed).

https://phabricator.endlessm.com/T14332