Add script to generate the json manifest
Since updating Firefox with so many language packs is complicated,
we now have a new script that will fetch the extra data details for
a given version of Firefox and its languages.

The script takes a while to run because in order to fetch the sha256
checksum we need to download all the mentioned assets. Still this
should be done only once and when there's an update to be done.

https://phabricator.endlessm.com/T14332