Update to latest NPAPI Flash
Adobe is now supporting the latest version for NPAPI.
Previously, for the past few years they had been only adding
security updates for the NPAPI build.

As part of this, the download path and file structure has changed,
eliminating the need to unzip an inner tar file (and significnatly
reducing the download file size, since we can now download just
the 64-bit binary we need).

https://phabricator.endlessm.com/T15713