Query the LKGV file in S3 for the flash plugin and enable updates
At the moment, Firefox was installing one version of Flash and sticking
to it forever, which was a reasonable approach considering that Adobe
was not releasing new versions of the NPAAPI plugin for Linux. However,
that has recently changed and Adobe is now releasing versions of this
plugin more regularly, so it makes sense not only to make sure it gets
installed but also that it gets regularly updated.

To implement this, we borrowed some code of the updater we already use
for chromium and do a similar check for new versions (every 7 days, max)
and update it locally if needed. Also, to prevent depending on OS upgrades
just to update the latest version available from Adobe (which tends to
remove the old one from their servers after a few days), we also poke
out LKGV file in S3 (which we can update separately) first, and only
if that file or the relevant fields are not found, we use the fallback
values hardcoded in the script.

Last, note that the code added here never calls exit_with_error(), which
is done on purpose so that a failure installing the flash plugin does
not prevent the script from installing the Java plugin, just afterwards.

https://phabricator.endlessm.com/T15713