Allow firefox to set itself as a default browser
It does it by talking to GConf over D-Bus, so we need to allow firefox
to talk to it.

That does not fully work yet, somehow firefox does not detect that it
is not a default browser anymore when I change this in
gnome-control-center. It probably tries to change wrong settings…