Install an icon
Earlier we were taking an icon from the git repo - there is a
browser/branding directory containing the icons. But since we simply
download a prebuilt firefox, we don't get the icon anymore. Also, the
original icon does not fix the endless theme, so now we are installing
the firefox icon the eos-shell-content.