Remove the branch field
The branch field has the specific meaning - it describes the branch of
the application being flatpakked. It is also used for getting the base
app version if it is not explicitly specified.

We do not use any base app here nor is "52.0.1esr" a name of a branch
in mozilla's gecko-dev repository.