Use the Rust Sdk Extension
Rust is popular enough that there's an Sdk extension that can be
used instead fo bundling and/or building Rust as a module.

The extension is distributed from flathub. It has branches for both
Freedesktop 1.6 and Freedesktop 18.08 based Sdks. The branch you
will need depends on the Sdk you are going to use. GNOME-sdk 3.26/28
are based upon freedesktop-sdk 1.6 while GNOME-sdk 3.30 will be based
upon Freedesktop-sdk 18.08.